Add 'order' as an additional (optional) argument for recursive_ols
This commit lets recursive_ols step through the observations in an
arbitrary order that can be specified as an additional argument.
The default is 1:n. This lets us implement the bootstraps by passing
a new index as an additional argument, rather than reordering the
x and y matrices themselves and passing the reordered data as an
argument, which should hopefully be faster.