Start to speed up recursive_ols code w/ explicit QR updating
This commit switches the recursive_ols functions to update beta-hat with
Givens transformations of the previous QR factorization, instead of
refactoring the whole thing recursively.

I still need to change the algorithm for the final R \ r in each estimate
to impose that R is an upper triangular matrix. Unfortunately, the R here
is a submatrix. There is likely to be a better approach to storage.