Change OLS recursive estimation to impose upper-triangular matrices where we can
This commit changes the OLS recursive estimation algorithm so that we
explicitly call the LAPACK algorithms for upper triangular matrices
in the inversion R \ r. Here we know that R is upper triangular
because it is comes from the QR decomposition (and is then multiplied
by Givens rotations, which preserve the upper triangular structure).